+++ powersync/src/sync/upload.rs
@@ -1,12 +1,13 @@
-use std::{collections::HashSet, sync::Arc};
+use std::{collections::HashSet, ops::ControlFlow, sync::Arc};
 
 use futures_lite::{
     FutureExt, StreamExt,
     future::{self, Boxed},
 };
 use log::{debug, info, warn};
-use rusqlite::{Connection, TransactionBehavior, params};
+use powersync_sqlite_nostd::{Destructor, ResultCode};
 
+use crate::db::connection::{SqliteConnection, TransactionGuard};
 use crate::db::watch::ListenerConfiguration;
 use crate::sync::coordinator::SyncCoordinator;
 use crate::{
@@ -59,7 +60,7 @@ impl UploadActor {
             .env
             .pool
             .update_notifiers()
-            .listen(ListenerConfiguration::if_matches(tables, false));
+            .listen(ListenerConfiguration::if_matches(tables, true));
         ConnectedUploadActor {
             connector,
             crud_stream: stream.map(|_| ()).boxed(),
@@ -158,55 +159,16 @@ impl UploadActor {
                     Self::state_transition_from_command_while_uploading(&self.commands, &self.db);
 
                 let upload_done = async {
-                    let (result, state) = result.await;
+                    let state = result.await;
+                    self.db
+                        .status
+                        .update(|s| s.set_upload_state(UploadStatus::Idle));
 
-                    match result {
-                        Ok(_) => {
-                            // It's possible that pending CRUD uploads were preventing data from
-                            // syncing. So now that that's completed, notify the download client in
-                            // case it needs to retry.
-                            if let Some(sync) = self.db.sync.upgrade() {
-                                sync.mark_crud_uploads_completed().await;
-                            }
-
-                            // Apart from that, the upload is done and we transition back into the
-                            // ready connected state to start the next iteration when needed.
-                            Some(UploadActorState::Connected(state))
-                        }
-                        Err(e) => {
-                            warn!("CRUD uploads failed, will retry, {e}");
-                            self.db
-                                .status
-                                .update(|s| s.set_upload_state(UploadStatus::Error(e)));
-                            let db = self.db.clone();
-
-                            Some(UploadActorState::WaitingForReconnect {
-                                timeout: async move {
-                                    db.sync_iteration_delay().await;
-                                    state
-                                }
-                                .boxed(),
-                            })
-                        }
-                    }
-                };
-
-                future::race(request, upload_done)
-                    .await
-                    .unwrap_or(old_state)
-            }
-            UploadActorState::WaitingForReconnect { ref mut timeout } => {
-                // Either the timeout expires, in which case we reconnect, or a disconnect is
-                // requested.
-                let request =
-                    Self::state_transition_from_command_while_uploading(&self.commands, &self.db);
-
-                let timeout_expired = async {
-                    let state = timeout.await;
+                    // The upload is done and we transition back into the  ready connected state to start the next iteration when needed.
                     Some(UploadActorState::Connected(state))
                 };
 
-                future::race(request, timeout_expired)
+                future::race(request, upload_done)
                     .await
                     .unwrap_or(old_state)
             }
@@ -222,9 +184,9 @@ impl UploadActor {
                     connector: state.connector.as_ref(),
                     db,
                 };
-                let result = upload.run().await;
+                upload.run().await;
 
-                (result, state)
+                state
             }
             .boxed(),
         }
@@ -234,12 +196,7 @@ impl UploadActor {
 enum UploadActorState {
     Idle,
     Connected(ConnectedUploadActor),
-    RunningUpload {
-        result: Boxed<(Result<(), PowerSyncError>, ConnectedUploadActor)>,
-    },
-    WaitingForReconnect {
-        timeout: Boxed<ConnectedUploadActor>,
-    },
+    RunningUpload { result: Boxed<ConnectedUploadActor> },
     Stopped,
 }
 
@@ -262,85 +219,122 @@ struct CrudUpload<'a> {
 }
 
 impl<'a> CrudUpload<'a> {
-    pub async fn run(&mut self) -> Result<(), PowerSyncError> {
+    pub async fn run(&mut self) {
         let mut last_item_id = None::<i64>;
 
-        while let Some(item) = self.oldest_crud_item_id().await? {
-            if last_item_id == Some(item) {
-                warn!("{}", Self::DUPLICATE_ITEM_WARNING);
-                return Err(PowerSyncError::argument_error(
-                    "Delaying due to previously encountered CRUD item.",
-                ));
+        // Invoke upload method on connector until there are no remaining CRUD items to upload.
+        loop {
+            match self.upload_step(&mut last_item_id).await {
+                Ok(ControlFlow::Break(_)) => break,
+                Ok(ControlFlow::Continue(_)) => continue,
+                Err(e) => {
+                    last_item_id = None;
+                    info!("CRUD uploads failed, will retry, {e}");
+
+                    self.db
+                        .status
+                        .update(|data| data.set_upload_state(UploadStatus::Error(e)));
+                    self.db.sync_iteration_delay().await;
+                }
+            }
+        }
+    }
+
+    async fn upload_step(
+        &mut self,
+        last_item_id: &mut Option<i64>,
+    ) -> Result<ControlFlow<()>, PowerSyncError> {
+        let Some(item) = self.oldest_crud_item_id().await? else {
+            // Uploading is completed, advance write checkpoint.
+            if let Some(advance_target) = self.sequence_for_checkpoint().await? {
+                let write_checkpoint = self.get_write_checkpoint().await?;
+                advance_target.complete(write_checkpoint, &self.db).await?;
             }
 
-            last_item_id = Some(item);
-            self.db
-                .status
-                .update(|data| data.set_upload_state(UploadStatus::Uploading));
-            self.connector.upload_data().await?;
+            // It's possible that pending CRUD uploads were preventing data from  syncing. So now
+            // that that's completed, notify the download client in case it needs to retry.
+            if let Some(sync) = self.db.sync.upgrade() {
+                sync.mark_crud_uploads_completed().await;
+            }
+
+            return Ok(ControlFlow::Break(()));
+        };
+
+        self.db
+            .status
+            .update(|data| data.set_upload_state(UploadStatus::Uploading));
+        if matches!(*last_item_id, Some(x) if x == item) {
+            warn!("{}", Self::DUPLICATE_ITEM_WARNING);
+            return Err(PowerSyncError::argument_error(
+                "Delaying due to previously encountered CRUD item.",
+            ));
         }
 
-        // Uploading is completed, advance write checkpoint.
-        if let Some(advance_target) = self.sequence_for_checkpoint().await? {
-            let write_checkpoint = self.get_write_checkpoint().await?;
-            advance_target.complete(write_checkpoint, &self.db).await?;
-        }
+        *last_item_id = Some(item);
+        self.connector.upload_data().await?;
 
-        Ok(())
+        Ok(ControlFlow::Continue(()))
     }
 
     async fn oldest_crud_item_id(&self) -> Result<Option<i64>, PowerSyncError> {
         let reader = self.db.reader().await?;
-        Self::read_oldest_crud_item_id(&reader)
+        Self::read_oldest_crud_item_id(reader.sqlite_connection())
     }
 
     async fn get_write_checkpoint(&self) -> Result<i64, PowerSyncError> {
         let client_id = {
             let reader = self.db.reader().await?;
-            let mut stmt = reader.prepare("SELECT powersync_client_id()")?;
-            let id: String = stmt.query_row(params![], |row| row.get(0))?;
-            id
+
+            let stmt = reader
+                .sqlite_connection()
+                .prepare("SELECT powersync_client_id()")?;
+            let ResultCode::ROW = stmt.step()? else {
+                panic!("Expected row"); // Can't happen, scalar select
+            };
+
+            stmt.column_text(0)?.to_string()
         };
 
         let credentials = self.connector.fetch_credentials().await?;
         write_checkpoint(&self.db, &client_id, credentials).await
     }
 
-    fn read_oldest_crud_item_id(conn: &Connection) -> Result<Option<i64>, PowerSyncError> {
-        let mut stmt = conn.prepare("SELECT id FROM ps_crud ORDER BY id LIMIT 1")?;
-        let mut rows = stmt.query(params![])?;
+    fn read_oldest_crud_item_id(conn: &SqliteConnection) -> Result<Option<i64>, PowerSyncError> {
+        let stmt = conn.prepare("SELECT id FROM ps_crud ORDER BY id LIMIT 1")?;
 
-        Ok(match rows.next()? {
-            None => None,
-            Some(row) => Some(row.get(0)?),
+        Ok(match stmt.step()? {
+            ResultCode::ROW => Some(stmt.column_int64(0)),
+            _ => None,
         })
     }
 
-    fn ps_crud_sequence(conn: &Connection) -> Result<Option<i64>, PowerSyncError> {
-        let mut seq_before = conn.prepare("SELECT seq FROM main.sqlite_sequence WHERE name = ?")?;
-        let mut seq_before = seq_before.query(params!["ps_crud"])?;
-        let Some(row) = seq_before.next()? else {
+    fn ps_crud_sequence(tx: &TransactionGuard) -> Result<Option<i64>, PowerSyncError> {
+        let seq_before = tx
+            .inner
+            .prepare("SELECT seq FROM main.sqlite_sequence WHERE name = ?")?;
+        seq_before.bind_text(1, "ps_crud", Destructor::STATIC)?;
+
+        let ResultCode::ROW = seq_before.step()? else {
             return Ok(None);
         };
 
-        Ok(row.get(0)?)
+        Ok(Some(seq_before.column_int64(0)))
     }
 
     async fn sequence_for_checkpoint(
         &self,
     ) -> Result<Option<PendingCheckpointRequest>, PowerSyncError> {
-        let reader = self.db.reader().await?;
-        {
-            let mut stmt =
-                reader.prepare("SELECT 1 FROM ps_buckets WHERE name = ? AND target_op = ?")?;
-            let mut rows = stmt.query(params!["$local", MAX_OP_ID])?;
-            if rows.next()?.is_none() {
-                // Nothing to update.
-                return Ok(None);
-            }
+        let mut reader = self.db.reader().await?;
+        let reader = reader.sqlite_connection_mut();
+        let read_tx = TransactionGuard::new(reader)?;
+
+        let current_target = InnerPowerSyncState::target_checkpoint_request_id(&read_tx, None)?;
+        if current_target != Some(MAX_OP_ID) {
+            // Nothing to update.
+            return Ok(None);
         }
 
-        let seq_before = Self::ps_crud_sequence(&reader)?;
+        let seq_before = Self::ps_crud_sequence(&read_tx)?;
         Ok(seq_before.map(|seq_before| PendingCheckpointRequest {
             crud_sequence: seq_before,
         }))
@@ -366,9 +360,9 @@ impl PendingCheckpointRequest {
         info!("Updating target to checkpoint {}", self.crud_sequence);
 
         let mut writer = db.writer().await?;
-        let writer = writer.transaction_with_behavior(TransactionBehavior::Immediate)?;
+        let writer = TransactionGuard::new(writer.sqlite_connection_mut())?;
 
-        if CrudUpload::read_oldest_crud_item_id(&writer)?.is_some() {
+        if CrudUpload::read_oldest_crud_item_id(writer.inner)?.is_some() {
             warn!("ps_crud is not empty, won't advance target");
             return Ok(());
         }
@@ -384,10 +378,7 @@ impl PendingCheckpointRequest {
             return Ok(());
         }
 
-        writer.execute(
-            "UPDATE ps_buckets SET target_op = ? WHERE name = ?",
-            params![op_id, "$local"],
-        )?;
+        InnerPowerSyncState::target_checkpoint_request_id(&writer, Some(op_id))?;
 
         writer.commit()?;
         Ok(())